Organize CLI into separate module
